multi-thread download, I don't understand how translation resource work so for now I left a static string

+++ app/src/main/java/eu/kanade/tachiyomi/data/download/anime/AnimeDownloader.kt
@@ -20,7 +20,9 @@ import eu.kanade.tachiyomi.data.cache.EpisodeCache
 import eu.kanade.tachiyomi.data.download.anime.model.AnimeDownload
 import eu.kanade.tachiyomi.data.library.anime.AnimeLibraryUpdateNotifier
 import eu.kanade.tachiyomi.data.notification.NotificationHandler
+import eu.kanade.tachiyomi.network.ProgressListener
 import eu.kanade.tachiyomi.util.storage.DiskUtil
+import eu.kanade.tachiyomi.util.storage.saveTo
 import eu.kanade.tachiyomi.util.storage.toFFmpegString
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -60,6 +62,8 @@ import uy.kohesive.injekt.injectLazy
 import java.io.File
 import java.util.Locale
 import kotlin.coroutines.cancellation.CancellationException
+import kotlin.math.min
+
 
 /**
  * This class is the one in charge of downloading episodes.
@@ -459,6 +463,7 @@ class AnimeDownloader(
                 tmpDir,
                 filename,
             )
+
             else -> {
                 if (preferences.useExternalDownloader().get() == download.changeDownloader) {
                     downloadVideo(video, download, tmpDir, filename)
@@ -503,12 +508,14 @@ class AnimeDownloader(
         video.status = Video.State.DOWNLOAD_IMAGE
         video.progress = 0
         var tries = 0
-
+        var forceSequential = false;
         // Define a suspend function to encapsulate the retry logic
         suspend fun attemptDownload(): UniFile {
             return try {
-                newDownload(video, download, tmpDir, filename)
+                newDownload(video, download, tmpDir, filename,forceSequential)
             } catch (e: Exception) {
+                // If the download failed, try again in sequential mode
+                forceSequential = true;
                 if (tries >= 2) throw e
                 tries++
                 delay((2 shl (tries - 1)) * 1000L)
@@ -651,11 +658,128 @@ class AnimeDownloader(
         return hours * 3600000L + minutes * 60000L + fullSeconds * 1000L + hundredths * 10L
     }
 
+    private suspend fun multiPartDownload(video: Video, download: AnimeDownload, tmpDir: UniFile, filename: String): UniFile {
+        //first we fetch the size of the video
+        val size: Long = download.source.getVideoSize(video, 3);
+        if (size == -1L) {
+            throw Exception("Could not get video size")
+        }
+        val partSize = size.div(preferences.numberOfThreads().get());
+        val partList = mutableListOf<UniFile>()
+        val rangeList = mutableListOf<Array<Long>>()
+        val partJobList = mutableListOf<Job>()
+        //create the parts as start byte and end byte
+        for (i in 0 until preferences.numberOfThreads().get()) {
+            val start = i * partSize
+            val end = if (i == preferences.numberOfThreads().get() - 1) {
+                size
+            } else {
+                (i + 1) * partSize - 1
+            }
+            rangeList.add(arrayOf(start, end))
+            val part = tmpDir.createFile("$filename.part$i.tmp")!!
+            partList.add(part)
+        }
+
+
+        //clear the tmp dir
+        tmpDir.listFiles()?.forEach { it.delete() }
+        var failed = false;
+        val totalProgresses = mutableListOf<Int>()
+        totalProgresses.addAll(List(preferences.numberOfThreads().get()) { 0 })
+
+
+        for (range in rangeList) {
+            val splitWeight = (range[1] - range[0]).toFloat() / size.toFloat()
+            //create a listener for each part, so we can update the progress generally
+            val listener = object : ProgressListener {
+                override fun update(bytesRead: Long, contentLength: Long, done: Boolean) {
+                    val progress = (((bytesRead * 100 / (range[1] - range[0])) * splitWeight).toInt())+1
+                    totalProgresses[rangeList.indexOf(range)] = progress
+                    video.progress = min( totalProgresses.reduce(Int::plus), 99)
+                }
+            }
+
+            //create a job for each part
+            partJobList.add(
+                launchIO {
+                    try {
+                        val response = download.source.getVideoChunk(video, range[0], range[1], listener)
+                        val file = tmpDir.findFile("$filename.part${rangeList.indexOf(range)}.tmp")
+                            ?: tmpDir.createFile("$filename.part${rangeList.indexOf(range)}.tmp")!!
+                        //try to open the file and append the bytes
+                        try {
+                            response.body.source().saveTo(file.openOutputStream(true))
+                        } catch (e: Exception) {
+                            response.close()
+                            failed = true
+                            throw e
+                        }
+                    } catch (e: Exception) {
+                        failed = true;
+                        throw e
+                    }
+                },
+            )
+
+
+        }
+
+
+        //wait for all parts to be downloaded
+        for (job in partJobList) {
+            if (failed) {
+                job.cancel()
+            } else {
+                job.join()
+            }
+        }
+            try {
+
+                val file0 = tmpDir.findFile("$filename.part0.tmp")
+                    ?: tmpDir.createFile("$filename.part0.tmp")!!
+                //merge all parts into file0
+                for (i in 1 until preferences.numberOfThreads().get()) {
+                    val part = tmpDir.findFile("$filename.part$i.tmp")
+                        ?: tmpDir.createFile("$filename.part$i.tmp")!!
+                    part.openInputStream().use { input ->
+                        file0.openOutputStream(true).use { output ->
+                            input.copyTo(output)
+                        }
+                    }
+                    part.delete()
+                }
+                file0.renameTo("$filename.mp4")
+
+            } catch (e: Exception) {
+
+                System.err.println(e.message)
+                throw e
+            }
+
+
+        if (failed) {
+
+            //delete all parts
+            for (i in 0 until preferences.numberOfThreads().get()) {
+                val part = tmpDir.findFile("$filename.part$i.tmp")
+                    ?: tmpDir.createFile("$filename.part$i.tmp")!!
+                part.delete()
+            }
+            throw Exception("Download failed")
+        }
+        val file = tmpDir.findFile("$filename.mp4") ?: throw Exception("Download failed");
+        return file
+
+
+    }
+
     private suspend fun newDownload(
         video: Video,
         download: AnimeDownload,
         tmpDir: UniFile,
         filename: String,
+        forceSequential: Boolean,
     ): UniFile {
         // Check if the download is paused before starting
         while (isPaused) {
@@ -665,38 +789,44 @@ class AnimeDownloader(
         if (isHls(video) || isMpd(video)) {
             return ffmpegDownload(video, download, tmpDir, filename)
         } else {
-            val response = download.source.getVideo(video)
-            val file = tmpDir.findFile("$filename.tmp") ?: tmpDir.createFile("$filename.tmp")!!
+            if (preferences.multithreadingDownload().get() &&!forceSequential) {
+                return multiPartDownload(video, download, tmpDir, filename)
+            } else {
 
-            // Write to file with pause/resume capability
-            try {
-                response.body.source().use { source ->
-                    file.openOutputStream(true).use { output ->
-                        val buffer = Buffer()
-                        var totalBytesRead = 0L
-                        var bytesRead: Int
-                        val bufferSize = 4 * 1024
-                        while (source.read(buffer, bufferSize.toLong()).also { bytesRead = it.toInt() } != -1L) {
-                            // Check if the download is paused, if so, wait
-                            while (isPaused) {
-                                delay(1000) // Wait for 1 second before checking again
+                val response = download.source.getVideo(video)
+                val file = tmpDir.findFile("$filename.tmp") ?: tmpDir.createFile("$filename.tmp")!!
+
+                // Write to file with pause/resume capability
+                try {
+                    response.body.source().use { source ->
+                        file.openOutputStream(true).use { output ->
+                            val buffer = Buffer()
+                            var totalBytesRead = 0L
+                            var bytesRead: Int
+                            val bufferSize = 4 * 1024
+                            while (source.read(buffer, bufferSize.toLong()).also { bytesRead = it.toInt() } != -1L) {
+                                // Check if the download is paused, if so, wait
+                                while (isPaused) {
+                                    delay(1000) // Wait for 1 second before checking again
+                                }
+
+                                // Write the bytes to the file
+                                output.write(bytesRead)
+                                totalBytesRead += bytesRead
+                                video.progress = (totalBytesRead * 100 / response.body.contentLength()).toInt()
+                                // Update progress here if needed
                             }
-
-                            // Write the bytes to the file
-                            output.write(bytesRead)
-                            totalBytesRead += bytesRead
-                            video.progress = (totalBytesRead * 100 / response.body.contentLength()).toInt()
-                            // Update progress here if needed
                         }
                     }
+
+                    // After download is complete, rename the file to its final name
+                    file.renameTo("$filename.mp4")
+                    return file
+                } catch (e: Exception) {
+                    response.close()
+                    if (!queueState.value.equals(download)) file.delete()
+                    throw e
                 }
-                // After download is complete, rename the file to its final name
-                file.renameTo("$filename.mp4")
-                return file
-            } catch (e: Exception) {
-                response.close()
-                if (!queueState.value.equals(download)) file.delete()
-                throw e
             }
         }
     }
